requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==4.0.4
django-rest-knox==4.2.0
djangorestframework==3.13.1
idna==3.3
Pillow==9.1.1
psycopg2-binary==2.9.3
pycparser==2.21
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
